more custom hooks, passenger path, some refactoring
