+++ GG_clone_backend/app/controllers/api/v1/drivers_controller.rb
@@ -33,17 +33,7 @@ class Api::V1::DriversController < ApplicationController
     else render json: @user.errors, status: :unprocessable_entity
     end
   end
-
-  def create_coordinates
-    driver = Driver.find(params[:id])
-    driver.latitude = params[:coordinates][:latitude]
-    driver.longitude = params[:coordinates][:longitude]
-    if driver.save
-      render json: { message: 'driver coordinates has been saved' }
-    else render json: { message: 'error! driver coordinates has not been saved' }, status: :unprocessable_entity
-    end
-  end
-
+  
   private
 
   def user_params
